refactor: Form 관련 클래스 폴더 정리, FormItem의 데이터 기본 타입을 byte[]로 변경 + 문자열 변환 기능 추가
