--- src/main/java/chain/record/FormItem.java
@@ -1,3 +1,0 @@
-package chain.record;
-
-public record FormItem(String filename, String data) {}
